requirements.txt
pandas==2.2.2
numpy==1.26.4
streamlit==1.37.0

requests==2.32.3
python-dotenv==1.0.1
pybaseball==2.2.7
